Changing the project type and adding a new test project and class
The project type has changed, and a new test project named "AdmxParser.Test" has been added to the AdmxParser.sln file. The AdmxDirectory.cs file added a new using statement and a block of code to load the available languages in the directory. In the AdmxParser.csproj file, the README.md file was reformatted and a new AssemblyAttribute for test projects was added. The AdmxDirectoryTest.cs and AdmxParser.Test.csproj files are new and contain code to test the functionality of the AdmxDirectory class and define the settings and dependencies of the test project, respectively. Finally, the WindowsLanguage.cs file is new and defines classes that provide information about the Windows language.