Added Group Policy Update functionality and fixed test code
Added new classes and methods to perform Group Policy updates. Added a new method in `AdmxExtensions.cs` to get the registry keys for the local machine and the current user, and added a reference to the `Microsoft.Win32.Registry` package in `AdmxParser.csproj` to do so. New classes to perform Group Policy updates have been added to `UpdatePolicy.cs` and `NativeMethods.cs`. In addition, a new test class to test these changes has been added to `UpdatePolicyTest.cs`.

In the existing code, the test class constructor in `AdmxDirectoryTest.cs` was removed, and the indentation of the namespace string in `Helpers.cs` was corrected.